refactor: split parsing and rendering
